Mounts experiment disks as ext4
To better withsand poweroff failures, try to mount the disks as ext4.
The files ping.disk and metdata.disk must be deleted before we can run this. 


This version alos adds data=journal and nodelalloc. 

This will impact perfomance (disabled delayed writes and force eveyrthing to be in sycn before commit) if we see a problem with slow disk writes maybe disable these options.