Add a recovery reboot script for the MF910
We sometimes see mifi's fail (or rather the connection with the node) with
errors like -32, -64 etc. Several of these are non recoverable and we have been
resorting to draining the battery in order to force a restart of the affected
mifi(s).

Draining the battery has turned out to be unecessary and overly complex. As we
discovered long time ago, the MF910 runs Android and comes with a fairly
complete version of busybox. It also runs a DHCP server (of course), so we know
which clients are connected. After some experimentation, we were able to log
into one broken mifi over wifi and run adb on it. Running reboot was enough to
restart USB (not always the case with embedded devices) and restore contact with
the node.

This commit adds a script for automatically rebooting the device when we detect
something is wrong. The script currently starts by checking for DHCP leases and
will use the first lease from a host with hostname that contains Monroe. This is
an OK assumption, there should never be more than one node connected to a mifi
at any point.

If no host is found for ten minutes, we reboot the mifi as we sometimes see that
mifis crash right after boot. Ten minutes should be more than enough for DHCP to
be started on the node. If DHCP has not been started after ten minutes,
something else is wrong and the state of the mifi is not important (or we can at
least safely reboot it).

After we have found the host, we start pinging it. We send one ping every fifth
second and wait one second for each reply. If we miss five consecutive replies,
we reboot the mifi. This is OK as the connection between the mifi and host is
close to lossless (at least based on our experience). Thus, five lost packets in
a row is a sign that something is wrong.